Move plot.py to separate file
